# subciber-seguridad-rest/src/main/webapp/swagger.json
{
    "swagger":"2.0",
    "info":{
        "version":"1.0",
        "title":"subciber-seguridad-rest"
    },
    "host":"",
    "basePath":"/subciber-seguridad-rest/",
    "schemes":[
        "http"
    ],
    "paths":{
...
        "/usuario":{
            "get":{
                "consumes":[
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "$ref":"#/definitions/ResponseUsuarioDto;>"
                        }
                    }
                }
            },
            "put":{
                "consumes":[
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "name":"body",
                        "in":"body",
                        "required":true,
                        "schema":{
                            "$ref":"#/definitions/UsuarioActualizacionDto"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "$ref":"#/definitions/UsuarioActualizacionDto;>"
                        }
                    }
                }
            },
            "post":{
                "consumes":[
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "name":"body",
                        "in":"body",
                        "required":true,
                        "schema":{
                            "$ref":"#/definitions/UsuarioDetalleDto"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "$ref":"#/definitions/UsuarioDetalleDto;>"
                        }
                    }
                }
            },
            "delete":{
                "consumes":[
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "name":"body",
                        "in":"body",
                        "required":true,
                        "schema":{
                            "$ref":"#/definitions/EliminarObjetoDto"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "$ref":"#/definitions/AuditResponseDto"
                        }
                    }
                }
            }
        },
...
    },
    "definitions":{
...
        "AuditResponseDto":{
            "properties":{
                "codigoRespuesta":{
                    "type":"integer"
                },
                "mensajeRespuesta":{
                    "type":"string"
                },
                "transaccionId":{
                    "type":"string"
                }
            }
        },
...
        "EliminarObjetoDto":{
            "properties":{
                "items":{
                    "type":"array",
                    "items":{
                        "type":"integer"
                    }
                }
            }
        },
...
        "ResponseUsuarioDto":{
            "properties":{
                "usuarios":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/VUsuario"
                    }
                }
            }
        },
        "ResponseUsuarioDto;>":{
            "properties":{
                "auditResponse":{
                    "$ref":"#/definitions/AuditResponseDto"
                },
                "objectResponse":{
                    "$ref":"#/definitions/ResponseUsuarioDto"
                }
            }
        },
...
        "UsuarioActualizacionDto":{
            "properties":{
                "datosUsuario":{
                    "$ref":"#/definitions/UsuarioDetalleDto"
                },
                "rol":{
                    "type":"array",
                    "items":{
                        "type":"integer"
                    }
                }
            }
        },
        "UsuarioActualizacionDto;>":{
            "properties":{
                "auditResponse":{
                    "$ref":"#/definitions/AuditResponseDto"
                },
                "objectResponse":{
                    "$ref":"#/definitions/UsuarioActualizacionDto"
                }
            }
        },
...
        "UsuarioDetalleDto":{
            "properties":{
                "apellido":{
                    "type":"string"
                },
                "clave":{
                    "type":"string"
                },
                "codigoUsuario":{
                    "type":"string"
                },
                "emailUsuario":{
                    "type":"string"
                },
                "imagenUsuario":{
                    "type":"string"
                },
                "nombre":{
                    "type":"string"
                },
                "rol":{
                    "type":"array",
                    "items":{
                        "type":"integer"
                    }
                },
                "tokenUsuario":{
                    "type":"string"
                },
                "usuario":{
                    "type":"string"
                },
                "usuarioId":{
                    "type":"integer"
                }
            }
        },
        "UsuarioDetalleDto;>":{
            "properties":{
                "auditResponse":{
                    "$ref":"#/definitions/AuditResponseDto"
                },
                "objectResponse":{
                    "$ref":"#/definitions/UsuarioDetalleDto"
                }
            }
        },
...
        "VUsuario":{
            "properties":{
                "apellido":{
                    "type":"string"
                },
                "aplicacionId":{
                    "type":"integer"
                },
                "codigo":{
                    "type":"string"
                },
                "email":{
                    "type":"string"
                },
                "estado":{
                    "type":"string"
                },
                "estadoId":{
                    "type":"integer"
                },
                "id":{
                    "type":"integer"
                },
                "imagen":{
                    "type":"string"
                },
                "nombre":{
                    "type":"string"
                },
                "usuario":{
                    "type":"string"
                }
            }
        },
...
    }
}